several simple fixes
