+++ testnet/stacks-node/src/run_loop/neon.rs
@@ -286,7 +286,8 @@ impl RunLoop {
     ) -> MockController {
         // Initialize and start the burnchain.
         let mut burnchain_controller =
-            MockController::new(self.config.clone(), coordinator_senders);
+        BurnchainController::new();
+        // BurnchainController::new(self.config.clone(), coordinator_senders);
 
         let burnchain_config = burnchain_controller.get_burnchain();
         let epochs = burnchain_controller.get_stacks_epochs();
